ktool: Add initial client support
ktool intends to be a command line tool for knot protocol testing.
It will expose options to allow registering, send and receive commands
from the server. Unix socket is the first transport supported for local
testing, other transports such as Bluetooth, TCP/IP and SPI may be
supported in the future.