gw: Add node driver
This patch adds an abstraction for radio technologies and socket based
connections. It intends to be a server abstraction to manage connection
based on the socket operations concept.

The initial use case is NRF24L01, SPI connection must be multiplexed for
each connected device (or pipe). Other uses include proxy for location
based service, BTLE, 802.15.4, and Wi-Fi.