gw: Rename 'recv' callback to 'get'
This patch renames 'recv' callback of the proto_ops 'driver' to
'get' in order to map the operation to the actual actions. For
HTTP/REST, the operation consist of sending a HTTP/GET with the
proper UUID and token set in the header to GET the device data.
For others IoT protocols, in theory all operations can be mapped
easily to GET/PUT operations.