tools: Add command line option to handle schemas
This patch extends ktool adding a new option '--schema' to allow
registering a data schema for a given KNOT device. The group option
'--uuid' defines device's UUID, and '--json' informs the data schema.
JSON format is used to provide a friendly schema representation. ktool
is responsible for translating to KNOT PDUs format and send to the
gateway.