tools: Add command line option to send data
This patch extends ktool adding a new option '--data' to allow
setting data for a given KNOT device. The group option '--uuid' defines
device's UUID, and '--json' informs the data to the sent. JSON format
is used to provide a friendly representation. ktool is responsible
for translating to KNOT PDUs format and send it to the gateway.