Add: Checks if the schema was already received before add to trust
Whenever the THING sends an schema, if it is not the last one,
it is added to the temporary trust list. However, sometimes the
THING sends it and does not receive the ack from the GW, what
causes retransmission. In that cases, the schema should not be
added to the list again.