settings: Disabling set 'id' to nobody
This patch temporarily disable set 'id' to nobody. Storage fails to
write in knotd configuration file since knotd does not have permission
to write in the file after setting 'id' to nobody. In order to allow
storage to write on knotd configuration file a temporarily approach is
to disable set nobody while studying a better one.