knotd: Set online property when create a device
When a new device is created the online property is set up to false as
default value. This behavior causes a bug if a proxied device is added.
The bug shows the online device as an offline device.

This patch set the online property when a device is created to fix the
bug.

Signed-off-by: Vitor Barros <vba@cesar.org.br>