Reorganize return values of cloud_receive_message
This patch reorganize the return values of cloud_receive_message.
In this function, the received message is consumed from AMQP queue
to be parsed. In case of an error on parse, the message is consumed,
but not used.
Theses changes make it more clear that the function returns true if the
message envelope was consumed or returns false otherwise.