Change names for a more suitable one
Change the name of function append_queue to append_update_queue because
refers to append on update_list.
Change parameter 'data' on on_update_cb and on_request_cb to 'list'.
Remove 'cb' postfix from on_update_cb and on_request_cb because the
'on' prefix is sufficient to represent a callback function.