Fix device unregister on schema rollback
This patch sends an unregister message only if the device is already
registered when the schema message is not received. Previously, when the
timeout occurred and the schema message did not arrive, a log message
was always sent. However, it is not always necessary, this should only
occur if the device is already connected.